Allow uploading new data
Create a landing page that offers an option to upload new data. On this
new page, have a form that allows choosing a file to upload.

Once the file is uploaded, store it, and display a helpful "Done"
page that also states the URL that is to be used for downloads in the
future.

For customization of the "Done" page, for now we have an extremely
simple "templating engine" that just use string replacements to replace
some placeholders. In the future this should probably be replaced by
something more elaborate.