Refactor code for file sharing
We have to separate handling the HTTP submission, and the writing to our
storage. This is in preparation for supporting automatic expiration of
uploaded data, which means we need to read multiple form fields. As the
order of the fields is not necessarily fixed, we first need to read
everything into our own data structure and only afterwards (if
all required information has been provided) write it to our storage.