Add Docker Container for Deployment
For easy deployment, add a Dockerfile that builds Teimeida using the
official Rust container and then (as part of the multi-stage build)
creates a very minimal runtime container containing only the binary and
the assets.

In order to use the container, port 8080 must be exposed and a data
volume must be mounted at `/data`, where all files will be stored.