Offer `config` Volume in Docker Container
To allow for the easy configuration of the Docker Container, let's
expose a `config` volume where the credentials file can be stored. In
fact, this should now work in such a way that the example config file is
stored in the volume if empty. I consider this a "good thing".